MengtingUI/main-tool/requirements.txt
pandas
pymupdf==1.18.12
dash-ag-grid==2.3.0
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-extensions==1.0.3
dash-html-components==2.0.0
dash-mantine-components==0.12.1
dash-table==5.0.0
dataclass-wizard==0.22.2
dataclasses-json==0.5.7
Flask==2.2.5
beautifulsoup4==4.12.2
Pillow==10.0.1
lxml
html5lib
jsonlines
google-cloud-aiplatform
opencv-python
pytesseract
nltk
autocorrect
unidecode
matplotlib
dash-iconify
dash-canvas
ipython
fire
PyAuth
dash-auth
pyyaml
gunicorn
